+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Added
+- The `resave/entries` command now supports a `--revisions` flag.
+
 ### Fixed
 - Fixed an error that occurred when initializing an element with custom field values that weren’t included in its field layout. ([#10622](https://github.com/craftcms/cms/issues/10622))
 - Fixed a PHP 8 compatibility bug.